phone_fingerprint/requirements.txt
# Phone-based Fingerprint Authentication System Requirements
# Core dependencies for fingerprint processing
numpy>=1.21.0
opencv-python>=4.8.0
Pillow>=9.0.0

# Additional dependencies for advanced processing
scipy>=1.9.0
scikit-learn>=1.0.0

# Web API integration (for phone communication)
flask>=2.3.0  # For web API integration
flask-cors>=4.0.0  # For CORS support
requests>=2.28.0  # For HTTP communication

# Optional: For other phone integration methods
# pyserial>=3.5  # For USB communication
# pybluez>=0.23  # For Bluetooth communication (Linux only)
# pyobjc-framework-CoreBluetooth>=9.0  # For Bluetooth on macOS
# pywin32>=305  # For Windows-specific features 